requirements.txt
absl-py==2.3.1
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==4.0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepface==0.0.95
deepgram-sdk==3.7.5
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
filelock==3.12.2
fire==0.7.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.8.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
limits==5.6.0
livekit==1.0.13
livekit-agents==1.2.16
livekit-api==1.0.5
livekit-blingfire==1.0.0
livekit-plugins-anam==1.2.16
livekit-plugins-cartesia==1.2.11
livekit-plugins-deepgram==1.2.11
livekit-plugins-elevenlabs==1.2.14
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.2.8
livekit-plugins-silero==1.2.11
livekit-plugins-turn-detector==1.2.11
livekit-protocol==1.0.6
lxml==6.0.0
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.99.6
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
priority==2.0.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
PyPDF2==3.0.1
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
Quart==0.20.0
quart-cors==0.8.0
regex==2025.7.34
reportlab==4.4.4
requests==2.32.4
retina-face==0.0.17
rich==13.9.4
safetensors==0.6.2
scipy==1.16.2
secure-smtplib==0.1.1
setuptools==80.9.0
silero==0.4.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
sympy==1.14.0
tbb==2021.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.0
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
